feat: Donwload application files
